Add tests for BlurInput
Summary:
I had to add React as a require in BlurInput to make this work, we
should probably go through and add those for the other components as
well. This might conflict with the current way we include React
components in webapp/Perseus though.

It would be nice to generalize the setting up of the jsdom etc. for
mounting components, so that's something to think about.

There seems to be a kind of weird bug with React where it will give an
error if I don't do the setting up of the window stuff beforehand:

```
Error: Invariant Violation: setEnabled(...): Cannot toggle event listening in a Worker thread. This is likely a bug in the framework.  Please report immediately.
```

We should ask about that, since I'm not sure it's a bug since we're
doing stuff with jsdom.

Test Plan: * make test

Reviewers: joel

Reviewed By: joel

Subscribers: jack

Differential Revision: http://phabricator.khanacademy.org/D8402