Add tests for ButtonGroup
Summary:
I considered doing something fancy to factor out the duplcation of props
but I decided just being explicit would probably be clearer and more
useful as an example-test.

I also had to add a ref to access the button, there might be a better
way to talk about the children of components? It might get a bit nasty
if we always have to go through the refs.

Test Plan: * make test

Reviewers: joel, brianmerlob

Reviewed By: joel, brianmerlob

Subscribers: jack

Differential Revision: http://phabricator.khanacademy.org/D8413