deprecate BackboneMixin
Move BackboneMixin to the bottom and give it a badge.

My biggest difficulty here was deciding whether it was more important
for baselines to be aligned or for the highlight to cover roughly the
line height of the header. In the end line height won. I'm not a
designer :D

Closes 11

Auditors: jack