Fixed LayeredComponentMixin example
Summary: The new layer is now fixed in the center of the page and includes
instructions for how to remove (i.e., to click again).

Test Plan: Ran examples page. Clicked 'Click me!' button multiple times (to toggle on and off),
scrolled around page to ensure that div remained fixed.