Upgrade katex to the newest version
Summary:
The katex interface changed, `katex.process` -> `katex.render`. Change
this in the `tex.js` interface. Also, include the fonts, etc. to make
the docs actually work if the expression is katex-renderable (we might
soon support the example expression).

Of note for future upgrades: I had to change the path that katex finds
the fonts at by changing `fonts.less` in katex to point to `../fonts`.

Test Plan:
 - Remove the `\nabla` from the tex example.
 - `make docs`
 - Serve the docs/index.html file
 - Make sure that the TeX example works in that file.

Reviewers: joel, alpert

Reviewed By: alpert

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D14268