Get rid of the other addons.
Summary:
It turns out the doing `require("react/addons")` is actually a
mutative operation that adds the addons to everywhere `require("react")`
is used as well, which means that if a test with
`require("react/addons")` ran before the i18n test, the module would
succeed even though it wasn't requiring the right thing. I "fixed" this
when I made it do `require("react/addons")`, but then this broke on
KA.org which only has a `require("react")`. I could have done some funky
hack but just getting rid of all of the addons seemed easier.

Test Plan: - `make test`

Reviewers: alpert

Reviewed By: alpert

Differential Revision: https://phabricator.khanacademy.org/D22123