Make `TeX` wait for MathJax to load before rendering.
Summary:
Before this, we depended on MathJax loading before we even
tried to render any `TeX` component. This was annoying, because we'd
rather not have to wait for MathJax to have fully loaded before we
render this.

This change uses the maybe-existent `Khan.mathJaxLoaded` Promise to wait
until MathJax is actually loaded before trying to use it. This is kinda
annoying that it depends on KA-specific code but :\

Test Plan:
 - Substate this into perseus, substate that into webapp
 - Try to load a perseus article with MathJax in it
 - See that the article renders, and then after a bit the MathJax pops
   in.
 - `make test`

Reviewers: alex

Reviewed By: alex

Subscribers: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D26123