Trigger all callbacks when rendering TeX via MathJax
Summary:
Previously, we were only triggering the `onRender` callback for, e.g., a single `Tex` component
rendered with MathJax, rather than for _every_ Tex component rendered with MathJax--since
for every `TeX` component beyond the first, we'd see that we were in the midst of processing a
batch of scripts, push our script onto the stack, and continue.

This diff modifies the script-processing logic to track each callback, along with the scripts that
are pushed to the queue, such that we can then trigger each callback in turn after batch-processing
the scripts.

On Asana @ https://app.asana.com/0/132409062806753/145312298229364.

Test Plan:
- Substate this change into Perseus; substate Perseus into webapp.
- Enable the XOM Manatee beta.
- Open up an article with multiple blocks of MathJax content ([link](http://localhost:8080/math/algebra/two-var-linear-equations/writing-slope-intercept-equations/a/writing-slope-intercept-equations)).
    - Verify that, after the initial render, all of the content fades in, rather than just the first block.

Reviewers: benkomalo, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D28342