Allow the Tooltip target container to have a custom style
Summary:
The target markup is wrapped in a `div` which is a problem if you want
the target markup to be inline with other text.  Add the ability to style
the target markup container to avoid this problem.

Use PropTypes.any

Test Plan:
Use a tooltip and include a `targetContainerStyle` to style the
markup found in the target section

Reviewers: riley

Reviewed By: riley

Differential Revision: https://phabricator.khanacademy.org/D32039