Fix up a number of issues in react-components.
Summary:
- Removed all lint errors
- Removed all usage of Underscore (this makes it so that react-components.js no longer depends on corelibs-legacy.js!)
- Upgrade to use the latest version of React
- Upgraded all dev dependencies to latest versions
- Fixed issues in react-live-editor and brought in that code:
https://github.com/Khan/react-live-editor/commit/9a805f85f816286e86ad9a461bb0479d1e96ab21

Test Plan:
- Confirmed that "make test" passes and runs without warnings.
- Confirmed that "make" runs.
- Confirmed that the docs/index.html page loads without errors or warnings *

* i18n.js seems to produce errors in the demo page, specifically:

"Invariant Violation: $_(...): A valid React element (or null) must be returned. You may have returned undefined, an array or some other invalid object."

It works fine in the tests, I'm not sure what's up! The implementation also seems to be the same as what we have in webapp, I'm assuming that it's an issue with the code editor and how the code is being dynamically compiled.

Reviewers: emily, alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D32383