react-components: add onResourceLoaded prop to TeX
Summary:
https://app.asana.com/0/27216215224639/221056683485381

Womp womp, if we measure a Perseus Sortable card before the KaTeX font is loaded, the browser makes incorrect guesses about how wide the characters will be, and we end up fixing the card's size too small and needing a horizontal scrollbar.

{F592837}

This change allows the caller to provide an `onResourceLoaded` callback that, in some browsers, will be called when any KaTeX font loads. This is a hint to the caller that the node's size may have changed, which gives Sortable the opportunity to re-measure.

We're punting this bug in browsers without `document.fonts` support, because, well, bleh. It's a minor visual bug, it seems to only happen on the mission dashboard when your first exercise is a Sortable (because otherwise a previous exercise probably loaded KaTeX fonts already), so I don't wanna sacrifice performance on all other browsers to fix an IE edge case :/

Test Plan:
In Perseus, update the Sortable component to remeasure the cards when `onResourceLoaded` fires. (diff incoming soon :D)
Confirm that, on the mission dashboard for Algebra, the "Ordering rational numbers" task no longer has horizontal scrollbars on its cards.
And `npm test` in the react-components repo to make sure I'm not crazy :)

Tested in Chrome 54 (fixed), mobile Chrome 54 (fixed), Firefox 50 (fixed), Safari 10 (fixed), Safari 7 (API unsupported), mobile Safari 10 (API supported, devtools is hard so didn't confirm fix), mobile Safari 9 (API unsupported), Edge 14 (API unsupported), IE 10 (API unsupported). In browsers where the API is supported, I confirmed that the cards are still sortable, and used DevTools to confirm that the widths are remeasured after KaTeX fonts load (and a couple refreshes to visually confirm). In browsers where the API is unsupported, I only confirmed that the cards are still sortable.

Reviewers: emily, aasmund, kevinb

Reviewed By: kevinb

Subscribers: lauren, kevinb

Differential Revision: https://phabricator.khanacademy.org/D32337