forward DragTarget props to root element
This fixes a regression in D32383: the Perseus `Editor` component expects to pass `children` and `className` props to the root `div` of `DragTarget`.
(We only noticed this issue recently, because we only recently updated Perseus to the latest version of `react-components` in D32410.)

We //could// explicitly pass down the props we expect, but, meh, it makes sense to attach arbitrary props to the root element, and I like this API better. It also doesn't require an update to the Perseus call site. If we decide we want a different API, though, we can make that change in a later revision.

Test Plan:
The Perseus One editor no longer crashes on initial render, yaaay >_<

Auditors: emily, john