Update the <TeX> component to the latest upstream KaTeX
Summary:
This diff updates docs/js/katex.js, docs/stylesheets/katex.css and
docs/stylesheets/fonts to KaTeX v0.9.0 from the upstream repo.

This diff also modifies the <TeX> component to use the new
colorIsTextColor option for backward-compatible handling of \color{}.

It adds a test to ensure that the option has the desired effect.
It also makes a number of updates to js/katex-a11y.js
to bring it back in sync with the latest KaTeX.

This is a change I made once before with a pre-release version of v0.9.0
but I had to revert that previous change. This time, I've just cherry-picked
the old version, and then updated the KaTeX files again.

Test Plan: - make test

Reviewers: emily, kevinb

Reviewed By: kevinb

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D42532