Don't continue to render MathJax if we re-render KaTeX quickly
Summary:
In the <TeX /> component, we have some code to correctly render the
math that is provided, switching between KaTeX and MathJax depending on whether
the math can be rendered with KaTeX or not.

Since MathJax is asynchronous, we sometimes will "queue" a render of some
MathJax math to happen as soon as possible. Sometimes, we'll end up
re-rendering in that time with some math that uses KaTeX. When that happens,
because we queued the render of the MathJax to happen async, it will continue
to happen, and we'll end up with both the MathJax and KaTeX rendered at the
same time!

This diff attempts to solve that problem by de-queueing the render of MathJax
math when the <TeX /> component is re-rendered and the previous queued render
hasn't happened yet.

Jira task: https://khanacademy.atlassian.net/browse/CP-173

Test Plan:
 - Visit http://localhost:8081/math/arithmetic/fraction-arithmetic/arith-review-fractions-on-the-number-line/e/fractions_on_the_number_line_1
 - Do a bunch of problems
 - See that we never end up with two bits of math next to each other

Reviewers: davidflanagan, kevinb

Reviewed By: davidflanagan

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D42653