Import katex through node_modules in react-components
Summary: Instead of assume `katex` is globally available, we will `require` it. Kake will look for it in webapp, which is why I've added an entry to `javascript/node_modules/katex` in D48636 — but in principle, this will work once this package has its dependencies resolved by Webpack.

Test Plan: See D48636

Reviewers: #webpack, john

Reviewed By: #webpack, john

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D48724