Fallback to MathJAX on any Katex error, not just parse errors
Summary:
Bugs CP-879 and CP-1742 are caused by a TypeError thrown by KaTeX.
We already fall back to MathJAX rendering when KaTeX throws a parse
error. But currently we rethrow other kinds of errors. This patch
changes that so we fallback to MathJAX for any kind of KaTeX error.

The underlying KaTeX error has been fixed upstream and will go away
when we upgrade webapp to use KaTeX 1.0. Since we've got the MathJAX
fallback, however, it seems worth using it in case other errors arise
in the future.

Test Plan:
Visit the following URL with this patch and verifies that it renders:
http://localhost:8081/math/k-8-grades/fr-v2-terminale-s/fr-v2-matrices-enseignement-de-spcialit/fr-v2-oprations-sur-les-matrices/a/multiplying-matrices-by-scalars?lang=fr

Reviewers: kevinb, michaelpolyak

Reviewed By: kevinb

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D50139